add chalk to package.json; move logic to cli index; remove how-to-disable guide
