+++ package.json
@@ -54,6 +54,7 @@
   },
   "dependencies": {
     "camelcase": "^6.3.0",
+    "chalk": "^5.0.1",
     "chokidar": "^3.5.3",
     "commander": "^9.2.0",
     "connect": "^3.7.0",
